Extract authorize middeware form prepareContext
