+++ lib/cartodb/controllers/analyses.js
@@ -1,11 +1,13 @@
 var PSQL = require('cartodb-psql');
 var cors = require('../middleware/cors');
 var userMiddleware = require('../middleware/user');
+const authorize = require('../middleware/authorize');
 const dbConnSetup = require('../middleware/db-conn-setup');
 
-function AnalysesController(prepareContext, pgConnection) {
+function AnalysesController(prepareContext, pgConnection, authApi) {
     this.prepareContext = prepareContext;
     this.pgConnection = pgConnection;
+    this.authApi = authApi;
 }
 
 module.exports = AnalysesController;
@@ -16,6 +18,7 @@ AnalysesController.prototype.register = function (app) {
         cors(),
         userMiddleware(),
         this.prepareContext,
+        authorize(this.authApi),
         dbConnSetup(this.pgConnection),
         createPGClient(),
         getDataFromQuery({ queryTemplate: catalogQueryTpl, key: 'catalog' }),
