+++ lib/cartodb/controllers/layergroup.js
@@ -3,6 +3,7 @@ const userMiddleware = require('../middleware/user');
 const allowQueryParams = require('../middleware/allow-query-params');
 const vectorError = require('../middleware/vector-error');
 const dbConnSetup = require('../middleware/db-conn-setup');
+const authorize = require('../middleware/authorize');
 const DataviewBackend = require('../backends/dataview');
 const AnalysisStatusBackend = require('../backends/analysis-status');
 const MapStoreMapConfigProvider = require('../models/mapconfig/provider/map-store-provider');
@@ -30,7 +31,7 @@ const SUPPORTED_FORMATS = {
  * @constructor
  */
 function LayergroupController(prepareContext, pgConnection, mapStore, tileBackend, previewBackend, attributesBackend,
-                              surrogateKeysCache, userLimitsApi, layergroupAffectedTables, analysisBackend) {
+                              surrogateKeysCache, userLimitsApi, layergroupAffectedTables, analysisBackend, authApi) {
     this.pgConnection = pgConnection;
     this.mapStore = mapStore;
     this.tileBackend = tileBackend;
@@ -44,6 +45,7 @@ function LayergroupController(prepareContext, pgConnection, mapStore, tileBacken
     this.analysisStatusBackend = new AnalysisStatusBackend();
 
     this.prepareContext = prepareContext;
+    this.authApi = authApi;
 }
 
 module.exports = LayergroupController;
@@ -56,6 +58,7 @@ LayergroupController.prototype.register = function(app) {
         cors(),
         userMiddleware(),
         this.prepareContext,
+        authorize(this.authApi),
         dbConnSetup(this.pgConnection),
         createMapStoreMapConfigProvider(this.mapStore, this.userLimitsApi),
         getTile(this.tileBackend, 'map_tile'),
@@ -76,6 +79,7 @@ LayergroupController.prototype.register = function(app) {
         cors(),
         userMiddleware(),
         this.prepareContext,
+        authorize(this.authApi),
         dbConnSetup(this.pgConnection),
         createMapStoreMapConfigProvider(this.mapStore, this.userLimitsApi),
         getTile(this.tileBackend, 'map_tile'),
@@ -97,6 +101,7 @@ LayergroupController.prototype.register = function(app) {
         cors(),
         userMiddleware(),
         this.prepareContext,
+        authorize(this.authApi),
         dbConnSetup(this.pgConnection),
         createMapStoreMapConfigProvider(this.mapStore, this.userLimitsApi),
         getTile(this.tileBackend, 'maplayer_tile'),
@@ -117,6 +122,7 @@ LayergroupController.prototype.register = function(app) {
         cors(),
         userMiddleware(),
         this.prepareContext,
+        authorize(this.authApi),
         dbConnSetup(this.pgConnection),
         createMapStoreMapConfigProvider(this.mapStore, this.userLimitsApi),
         getFeatureAttributes(this.attributesBackend),
@@ -136,6 +142,7 @@ LayergroupController.prototype.register = function(app) {
         userMiddleware(),
         allowQueryParams(['layer']),
         this.prepareContext,
+        authorize(this.authApi),
         dbConnSetup(this.pgConnection),
         createMapStoreMapConfigProvider(this.mapStore, this.userLimitsApi, forcedFormat),
         getPreviewImageByCenter(this.previewBackend),
@@ -153,6 +160,7 @@ LayergroupController.prototype.register = function(app) {
         userMiddleware(),
         allowQueryParams(['layer']),
         this.prepareContext,
+        authorize(this.authApi),
         dbConnSetup(this.pgConnection),
         createMapStoreMapConfigProvider(this.mapStore, this.userLimitsApi, forcedFormat),
         getPreviewImageByBoundingBox(this.previewBackend),
@@ -188,6 +196,7 @@ LayergroupController.prototype.register = function(app) {
         userMiddleware(),
         allowQueryParams(allowedDataviewQueryParams),
         this.prepareContext,
+        authorize(this.authApi),
         dbConnSetup(this.pgConnection),
         createMapStoreMapConfigProvider(this.mapStore, this.userLimitsApi),
         getDataview(this.dataviewBackend),
@@ -205,6 +214,7 @@ LayergroupController.prototype.register = function(app) {
         userMiddleware(),
         allowQueryParams(allowedDataviewQueryParams),
         this.prepareContext,
+        authorize(this.authApi),
         dbConnSetup(this.pgConnection),
         createMapStoreMapConfigProvider(this.mapStore, this.userLimitsApi),
         getDataview(this.dataviewBackend),
@@ -222,6 +232,7 @@ LayergroupController.prototype.register = function(app) {
         userMiddleware(),
         allowQueryParams(allowedDataviewQueryParams),
         this.prepareContext,
+        authorize(this.authApi),
         dbConnSetup(this.pgConnection),
         createMapStoreMapConfigProvider(this.mapStore, this.userLimitsApi),
         dataviewSearch(this.dataviewBackend),
@@ -239,6 +250,7 @@ LayergroupController.prototype.register = function(app) {
         userMiddleware(),
         allowQueryParams(allowedDataviewQueryParams),
         this.prepareContext,
+        authorize(this.authApi),
         dbConnSetup(this.pgConnection),
         createMapStoreMapConfigProvider(this.mapStore, this.userLimitsApi),
         dataviewSearch(this.dataviewBackend),
@@ -255,6 +267,7 @@ LayergroupController.prototype.register = function(app) {
         cors(),
         userMiddleware(),
         this.prepareContext,
+        authorize(this.authApi),
         dbConnSetup(this.pgConnection),
         analysisNodeStatus(this.analysisStatusBackend),
         sendResponse()
