+++ lib/cartodb/controllers/map.js
@@ -8,6 +8,7 @@ const cors = require('../middleware/cors');
 const userMiddleware = require('../middleware/user');
 const allowQueryParams = require('../middleware/allow-query-params');
 const dbConnSetup = require('../middleware/db-conn-setup');
+const authorize = require('../middleware/authorize');
 const NamedMapsCacheEntry = require('../cache/model/named_maps_entry');
 const NamedMapMapConfigProvider = require('../models/mapconfig/provider/named-map-provider');
 const CreateLayergroupMapConfigProvider = require('../models/mapconfig/provider/create-layergroup-provider');
@@ -28,7 +29,7 @@ const LayergroupMetadata = require('../utils/layergroup-metadata');
  */
 function MapController(prepareContext, pgConnection, templateMaps, mapBackend, metadataBackend,
                        surrogateKeysCache, userLimitsApi, layergroupAffectedTables, mapConfigAdapter,
-                       statsBackend) {
+                       statsBackend, authApi) {
     this.pgConnection = pgConnection;
     this.templateMaps = templateMaps;
     this.mapBackend = mapBackend;
@@ -43,6 +44,7 @@ function MapController(prepareContext, pgConnection, templateMaps, mapBackend, m
 
     this.statsBackend = statsBackend;
     this.prepareContext = prepareContext;
+    this.authApi = authApi;
 }
 
 module.exports = MapController;
@@ -70,6 +72,7 @@ MapController.prototype.composeCreateMapMiddleware = function (useTemplate = fal
         userMiddleware(),
         allowQueryParams(['aggregation']),
         this.prepareContext,
+        authorize(this.authApi),
         dbConnSetup(this.pgConnection),
         initProfiler(isTemplateInstantiation),
         checkJsonContentType(),
