+++ lib/cartodb/controllers/named_maps.js
@@ -2,6 +2,7 @@ const NamedMapsCacheEntry = require('../cache/model/named_maps_entry');
 const cors = require('../middleware/cors');
 const userMiddleware = require('../middleware/user');
 const dbConnSetup = require('../middleware/db-conn-setup');
+const authorize = require('../middleware/authorize');
 const allowQueryParams = require('../middleware/allow-query-params');
 const vectorError = require('../middleware/vector-error');
 
@@ -29,7 +30,7 @@ function getRequestParams(locals) {
 }
 
 function NamedMapsController(prepareContext, namedMapProviderCache, tileBackend, previewBackend,
-                             surrogateKeysCache, tablesExtentApi, metadataBackend, pgConnection) {
+                             surrogateKeysCache, tablesExtentApi, metadataBackend, pgConnection, authApi) {
     this.namedMapProviderCache = namedMapProviderCache;
     this.tileBackend = tileBackend;
     this.previewBackend = previewBackend;
@@ -38,6 +39,7 @@ function NamedMapsController(prepareContext, namedMapProviderCache, tileBackend,
     this.metadataBackend = metadataBackend;
     this.prepareContext = prepareContext;
     this.pgConnection = pgConnection;
+    this.authApi = authApi;
 }
 
 module.exports = NamedMapsController;
@@ -50,6 +52,7 @@ NamedMapsController.prototype.register = function(app) {
         cors(),
         userMiddleware(),
         this.prepareContext,
+        authorize(this.authApi),
         dbConnSetup(this.pgConnection),
         getNamedMapProvider({
             namedMapProviderCache: this.namedMapProviderCache,
@@ -75,6 +78,7 @@ NamedMapsController.prototype.register = function(app) {
         userMiddleware(),
         allowQueryParams(['layer', 'zoom', 'lon', 'lat', 'bbox']),
         this.prepareContext,
+        authorize(this.authApi),
         dbConnSetup(this.pgConnection),
         getNamedMapProvider({
             namedMapProviderCache: this.namedMapProviderCache,
