+++ lib/cartodb/middleware/authorize.js
