+++ lib/cartodb/middleware/context/index.js
@@ -2,14 +2,12 @@ const locals = require('./locals');
 const cleanUpQueryParams = require('./clean-up-query-params');
 const layergroupToken = require('./layergroup-token');
 const credentials = require('./credentials');
-const authorize = require('./authorize');
 
-module.exports = function prepareContextMiddleware(authApi) {
+module.exports = function prepareContextMiddleware() {
     return [
         locals(),
         cleanUpQueryParams(),
         layergroupToken(),
         credentials(),
-        authorize(authApi)
     ];
 };
