+++ lib/cartodb/server.js
@@ -217,7 +217,7 @@ module.exports = function(serverOptions) {
 
     const prepareContext = typeof serverOptions.req2params === 'function' ?
         serverOptions.req2params :
-        prepareContextMiddleware(authApi);
+        prepareContextMiddleware();
 
     /*******************************************************************************************************************
      * Routing
@@ -233,7 +233,8 @@ module.exports = function(serverOptions) {
         surrogateKeysCache,
         userLimitsApi,
         layergroupAffectedTablesCache,
-        analysisBackend
+        analysisBackend,
+        authApi
     ).register(app);
 
     new controller.Map(
@@ -246,7 +247,8 @@ module.exports = function(serverOptions) {
         userLimitsApi,
         layergroupAffectedTablesCache,
         mapConfigAdapter,
-        statsBackend
+        statsBackend,
+        authApi
     ).register(app);
 
     new controller.NamedMaps(
@@ -257,12 +259,13 @@ module.exports = function(serverOptions) {
         surrogateKeysCache,
         tablesExtentApi,
         metadataBackend,
-        pgConnection
+        pgConnection,
+        authApi
     ).register(app);
 
     new controller.NamedMapsAdmin(authApi, templateMaps).register(app);
 
-    new controller.Analyses(prepareContext, pgConnection).register(app);
+    new controller.Analyses(prepareContext, pgConnection, authApi).register(app);
 
     new controller.ServerInfo(versions).register(app);
